Adding and deleting products

+++ server/src/instrument/instrument.controller.ts
@@ -22,7 +22,7 @@ import { JwtAuthGuard } from "../authentication/jwt-auth.guard";
 import { Role } from "src/authorization/role.enum";
 import { FileInterceptor } from "@nestjs/platform-express";
 import * as fs from "fs";
-import { IMAGES_ASSETS_PATH } from "../consts/images.consts";
+import { DEFAULT_IMAGE_FILE_NAME, IMAGES_ASSETS_PATH } from "../consts/images.consts";
 
 @Controller('instrument')
 export class InstrumentController {
@@ -56,8 +56,10 @@ export class InstrumentController {
   @UseGuards(JwtAuthGuard, RolesGuard)
   async deleteInstrument(@Request() req: any, @Param('id') id: string) {
     const instrumentToDelete = await this.instrumentService.getInstrumentByID(id);
+    const photoToDelete = instrumentToDelete.photoUrl;
     return this.instrumentService.deleteInstrument(id).then(() => {
-      fs.unlinkSync(IMAGES_ASSETS_PATH + instrumentToDelete.photoUrl);
+      if (photoToDelete !== DEFAULT_IMAGE_FILE_NAME)
+        fs.unlinkSync(IMAGES_ASSETS_PATH + instrumentToDelete.photoUrl);
     })
   }
